Add sword affix pools and loot generator
Introduced ScriptableObjects for sword prefix and suffix affix pools, and added a LootGenerator script. Updated related item, affix, and loot management scripts to support new affix pool structure. Also made minor updates to enemy, inventory, and database scripts to integrate the new loot and affix systems.

+++ Assets/Scripts/Equipment/Items/ItemInstanceSO.cs
@@ -10,7 +10,7 @@ public struct ItemAffix
 [Serializable]
 public class ItemInstance
 {
-    public int instanceId;   //unique ID for this specific item instance
+    public long instanceId;   //unique ID for this specific item instance
     public int baseId;       //ID referring to the base item type
     public ItemRarity rarity;
     public int itemLevel;
